// packages/backend/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  ActionRetrier: typeof ActionRetrier;
  IndexData: typeof IndexData;
  ai: typeof ai;
  article: typeof article;
  auth: typeof auth;
  bannerMessages: typeof bannerMessages;
  brokers: typeof brokers;
  chat: typeof chat;
  company: typeof company;
  companyNames: typeof companyNames;
  companyNews: typeof companyNews;
  count: typeof count;
  crons: typeof crons;
  dashboard: typeof dashboard;
  "data/brokers": typeof data_brokers;
  "data/dp": typeof data_dp;
  "data/model": typeof data_model;
  email: typeof email;
  emails: typeof emails;
  exchangeMessages: typeof exchangeMessages;
  healthCheck: typeof healthCheck;
  http: typeof http;
  image: typeof image;
  indexChart: typeof indexChart;
  indexNames: typeof indexNames;
  internal: typeof internal_;
  ipo: typeof ipo;
  marketDepth: typeof marketDepth;
  marketStatus: typeof marketStatus;
  nepsePredictions: typeof nepsePredictions;
  news: typeof news;
  newsSites: typeof newsSites;
  notification: typeof notification;
  ohlc: typeof ohlc;
  ratelimit: typeof ratelimit;
  retrier: typeof retrier;
  rooms: typeof rooms;
  screenshot: typeof screenshot;
  seed: typeof seed;
  supplyDemand: typeof supplyDemand;
  triggers: typeof triggers;
  userLatency: typeof userLatency;
  users: typeof users;
  "utils/types": typeof utils_types;
  "utils/utils": typeof utils_utils;
  version: typeof version;
  webPush: typeof webPush;
  webSuggestions: typeof webSuggestions;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;